Implemented DomainNameResolver wrapper + test
File dns-resolver.h in the stout/libuv folder is for wrapping
uv_getaddrinfo(...) function in libuv library. This function
is for the asynchronous computation of the Internet address
with given node and service.
File dns-resolver.cc in the test/libuv folder is just for
testing logic which is implemented in stout/libuv/dns-resolver.h.