Change the pattern of some test cases
Previously in some tests we used a negative handler that 'should be' always invalid.
Those tests fail on Windows after a large amount of iterations.

This commit changes those tests so that every handler used in those cases is guaranteed
to be marked as invalid by closing it before the tested function.