Refactor streaming test
Before this commit, `test/streaming.cc` was a single large monolithic test that would not
allow us to test multiple different client behaviors easily. This PR leaves the test
functionally identical, but factors out the code that produces the client behavior
(which the next commit will introduce various versions of) from the framework of the test
(which will not change). It also splits `streaming.cc` into a separate cc_test target, so
that it can be built and tested in isolation more easily, for faster write/test-iteration.

As a bonus, adds a `.gitignore` file to this repo.