Only call `stream_->Finish()` after BOTH reads AND writes are done.
As documented in #12, we were calling `Finish()` too early - we shouldn't
call it before BOTH reads AND writes are done.

This PR corrects that problem.

It increases the number of times we lock a Mutex, but the increase is O(1): we
lock once more at start-time, and once more for both writes-finish and
reads-finish time. Small price to pay for avoiding persnickety bugs that come
with not using your APIs as you're supposed to. :-D

Fixes #12