Use uv_check_t instead of uv_prepare_t to simplify running the event loop
Currently 'EventLoop::Run()' needs to check if any callbacks have been
submitted to the event loop in order to determine if we should
continue executing the event loop or if we're all done. By using
'uv_check_t' instead of 'uv_prepare_t' we _check_ for the callbacks
after the event loop has completed it's firing timers, performing
pending work, etc so we can be sure that there won't be any other
callbacks submitted.