Ensure libuv functions are only used from within event loop
The general expectation for libuv functions is that they are only
called if the event loop is not running or from the event loop thread
in order to guarantee that two threads are not simultaneously
modifying the event loop. Thus, the safest strategy is to always call
libuv functions from within the event loop.