Introduce `DoAll()` implementation
Basically `DoAll()` takes a variadic list of eventuals, terminate them, store
them in a `std::tuple`, call `Start()` on all them, wait for them all to be
done, and then continue with a tuple of results from each of the eventuals
passed in. Each result may be either a result from eventual or `std::exception`.

Also added some tests.