Refactor interrupt semantics for 'Eventual', 'Loop', and 'Stream'
To simplify writing the "start" callback for 'Eventual', 'Loop', and
'Stream' we only install the interrupt handler _after_ invoking the
callback. Programmers still need to worry about the race of an
interrupt occurring at the same time that what ever they are waiting
for occurs, but at least they don't need to worry about any races
while they are doing any set up for what ever it is that they are
waiting for.

Note that we also removed interrupt handling from 'Terminal' because
if we've gotten to the 'Terminal' then we won't be waiting so there's
no reason to have an interrupt handler!