Add 'Client::Context()' to simplify 'Client::Call()'
Instead of requiring users of the client API to have to store their
own 'ClientContext' we provide an overload of 'Call()' which includes
a context as part of the eventual pipeline so that its address will be
stable. This allows us to remove declaring 'ClientContext' in all of
the tests that use a default context.