Introduce 'WaitForSignal()'
The current helpers for signal handling are relatively low-level as
they were designed to wrap the low-level 'uv_signal*()' family of
functions. In practice we often want to be able to specify multiple
signals and find out which signal was raised; this commit introduces
the helper 'WaitForSignal()' which does just that.