Refactor 'Expected' to be similar to 'Task'
Instead of 'Expected<T>' you do 'Expected::Of<T>' similar to
'Task::Of<T>'.

In addition, you can now use 'Expected(t)' to construct an
'Expected<decltype(t)>' without needing to specify the type as well as
do 'Unexpected(error)' without needing to specify the type and the
resulting types can be converted into the necessary 'Expected::Of<U>'
implicitly.