Reorder eventuals continuation 'k_' to be last
By always storing 'k_' as the _last_ member it will be destructed
_first_ and thus we won't have any use-after-delete issues during
destruction of 'k_' if it holds any references or pointers to any (or
within any) of the previous continuations.

This fixes 3rdparty/eventuals-grpc#71 and allows us to turn on
building with 'asan' on this repository as well as eventuals-grpc.