Add Dazel configuration
Before this PR the `eventuals` repo could only build if you had (the
right version of) `bazel` installed on your machine. To be more hermetic
we'd like to avoid this hope-you-installed-it-right type situations. In
fact, our GitHub Codespaces don't (currently) install `bazel` at all!

This PR adds a symlink to our standard `Dockerfile.dazel` in
the `dev-tools` subrepo, allowing us to use `dazel` to build in a
repeatable way. Dazel will, unfortunately, look for a `WORKSPACE` file
instead of a `WORKSPACE.dazel` file, so we add a symlink to that also.