Introduce initial support for "static" errors
Previously there was no way to know, or track, what errors might
propagate through a an eventual pipeline. This commit introduces the
initial components necessary to do that.

Next step is to add 'ErrorsFrom' to all eventual types and then update
'Composed' to not use 'ErrorsFromComposed' but instead assume that all
eventuals that are composed 'HaveErrorsFrom'.