Update from gRPC 1.40.0 to 1.45.0-pre1
This pulled in a newer version of rules_python (through an implicit
dependency via gRPC, which I don't like) which required porting off of
the `pip_import` rule which was deprecated in release 0.1.0 in Oct 2020
(!): https://github.com/bazelbuild/rules_python/releases/tag/0.1.0 and
deleted in release 0.6.0 in Jan 2022:
https://github.com/bazelbuild/rules_python/releases/tag/0.6.0.

This removes the need to depend on a 1-year old absl library (!).

It also unbreaks bazel 5.0: fixes https://github.com/3rdparty/eventuals-grpc/issues/65.