Use callables instead of eventuals for 'If.yes()' and 'If.no()' (#320)
Because most folks are not as used to expression based langauges they
run into issues where they end up doing things in the same eventuals
pipeline expression that they weren't expecting because they were
thinking like it was multiple statements. This is true of 'If.yes()'
and 'If.no()'. By making these take callables (e.g., a lambda) instead
then they are forced to treat the bodies of each not as expressions
and thus will be required to think (a little more) about how to
'std::move()' or copy values.