Ensure every 'Task' has a 'Scheduler::Context' when used as a continuation
When calling 'Start(...)', 'Fail(...)', 'Stop(...)' on a 'Task' as
though it's an already built _continuation_ we need to ensure that it
has it's own 'Scheduler::Context'. We were already doing this manually
in 'eventuals/grpc' but every task should always have it's own
context. This allows us to remove the now unnecessary context creation
in 'eventuals/grpc'.