Introduce second constructor for grpc::Client
Before this PR the only way to use a grpc::Client is to give it the
target and the channel credentials, and to have it then create the
actual gRPC channel. We've now found use cases where we want to create
the channel ourselves.

This PR introduces an additional constructor that allows us to create
a grpc::Client with an channel that we create ourselves.