Introduce 'TestingCompletionThreadPool'
Often when writing tests with gRPC you want to deterministically
ensure some ordering of events to test a particular code path. You can
do so by using a 'TestingCompletionThreadPool' which only has a single
thread that can be paused so as to ensure that no computation takes
place until a particular time or you can wait on it until there is no
more work to be done.